requirements.txt
# Pytorch
transformers
pytorch-lightning
torchmetrics
